RDMA/irdma: Set irdma_cq cq_num field during CQ create
The driver maintains a CQ table that is used to ensure that
a CQ is still valid when processing CQ related AEs.
When a CQ is destroyed, the table entry is cleared, using
irdma_cq.cq_num as the index. This field was never being set,
so it was just always clearing out entry 0.

Additionally, the cq_num field size was increased to accommodate
HW supporting more than 64K CQs.

Signed-off-by: Jacob Moroni <jmoroni@google.com>