irdma: Remove a NOP wait_event() in modify_qp_roce()
Remove a NOP wait_event() in modify_qp_roce() which is
relevant for iWARP and likely a copy and paste artifact for RoCEv2.
The wait event is for sending a reset on a TCP connection, after the
reset has been requested in modify_qp(), which occurs only in iWarp mode.

Signed-off-by: Tatyana Nikolova <tatyana.e.nikolova@intel.com>