irdma: Update ibqp state to error if already in error state
In modify_qp() update ibqp state to error if the irdma QP is already in
error state, otherwise the ibqp state which is visible to the consumer
app remains stale.

This issue is reported by a customer running iWARP where their app
encounters an AE which changes the irdma QP state to error. Next the app
calls modify_qp() to error which returns success without changing the
ibqp state and when the app queries the ibqp state, the state is still RTS.

Signed-off-by: Tatyana Nikolova <tatyana.e.nikolova@intel.com>